Add Response trait to dry code.
