+++ src/Illuminate/Http/JsonResponse.php
@@ -5,6 +5,8 @@ use Illuminate\Support\Contracts\JsonableInterface;
 
 class JsonResponse extends \Symfony\Component\HttpFoundation\JsonResponse {
 
+	use ResponseTrait;
+
 	/**
 	 * Get the json_decoded data from the response
 	 *
@@ -26,34 +28,6 @@ class JsonResponse extends \Symfony\Component\HttpFoundation\JsonResponse {
 		$this->data = $data instanceof JsonableInterface ? $data->toJson() : json_encode($data);
 
 		return $this->update();
-	}
-
-	/**
-	 * Set a header on the Response.
-	 *
-	 * @param  string  $key
-	 * @param  string  $value
-	 * @param  bool    $replace
-	 * @return \Illuminate\Http\Response
-	 */
-	public function header($key, $value, $replace = true)
-	{
-		$this->headers->set($key, $value, $replace);
-
-		return $this;
-	}
-
-	/**
-	 * Add a cookie to the response.
-	 *
-	 * @param  \Symfony\Component\HttpFoundation\Cookie  $cookie
-	 * @return \Illuminate\Http\Response
-	 */
-	public function withCookie(Cookie $cookie)
-	{
-		$this->headers->setCookie($cookie);
-
-		return $this;
 	}
 
 }
