+++ src/Illuminate/Http/ResponseTrait.php
@@ -1,0 +1,33 @@
+<?php namespace Illuminate\Http;
+
+trait ResponseTrait {
+
+	/**
+	 * Set a header on the Response.
+	 *
+	 * @param  string  $key
+	 * @param  string  $value
+	 * @param  bool    $replace
+	 * @return \Illuminate\Http\Response
+	 */
+	public function header($key, $value, $replace = true)
+	{
+		$this->headers->set($key, $value, $replace);
+
+		return $this;
+	}
+
+	/**
+	 * Add a cookie to the response.
+	 *
+	 * @param  \Symfony\Component\HttpFoundation\Cookie  $cookie
+	 * @return \Illuminate\Http\Response
+	 */
+	public function withCookie(Cookie $cookie)
+	{
+		$this->headers->setCookie($cookie);
+
+		return $this;
+	}
+
+}
